LayoutTests/fast/canvas/canvas-fillPath-pattern-shadow-expected.txt

Ensure correct behavior of canvas with fillPath using a pattern fillStyle and a shadow

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Verifying alpha shadow...
PASS data[0] is 0
PASS data[1] is 0
PASS data[2] is 0
PASS data[3] is 0
PASS data[0] is 255
PASS data[1] is 0
PASS data[2] is 0
PASS data[3] is around 64
PASS data[0] is 255
PASS data[1] is 0
PASS data[2] is 0
PASS data[3] is around 64
PASS data[0] is 255
PASS data[1] is 0
PASS data[2] is 0
PASS data[3] is around 64
PASS data[0] is 255
PASS data[1] is 0
PASS data[2] is 0
PASS data[3] is around 64

Verifying blurry shadow...
PASS data[0] is 0
PASS data[1] is 0
PASS data[2] is 0
PASS data[3] is 0
PASS data[0] is 255
PASS data[1] is 0
PASS data[2] is 0
PASS data[3] is around 31
PASS data[0] is 255
PASS data[1] is 0
PASS data[2] is 0
PASS data[3] is around 31
PASS data[0] is 255
PASS data[1] is 0
PASS data[2] is 0
PASS data[3] is around 31
PASS data[0] is 255
PASS data[1] is 0
PASS data[2] is 0
PASS data[3] is around 31

Verifying rotated alpha shadow...
PASS data[0] is 0
PASS data[1] is 0
PASS data[2] is 0
PASS data[3] is 0
PASS data[0] is 255
PASS data[1] is 0
PASS data[2] is 0
PASS data[3] is around 64
PASS data[0] is 255
PASS data[1] is 0
PASS data[2] is 0
PASS data[3] is around 64
PASS data[0] is 255
PASS data[1] is 0
PASS data[2] is 0
PASS data[3] is around 64
PASS data[0] is 255
PASS data[1] is 0
PASS data[2] is 0
PASS data[3] is around 64

Verifying rotated blurry shadow...
PASS data[0] is 0
PASS data[1] is 0
PASS data[2] is 0
PASS data[3] is 0
PASS data[0] is 255
PASS data[1] is 0
PASS data[2] is 0
PASS data[3] is around 31
PASS data[0] is 255
PASS data[1] is 0
PASS data[2] is 0
PASS data[3] is around 31
PASS data[0] is 255
PASS data[1] is 0
PASS data[2] is 0
PASS data[3] is around 31
PASS data[0] is 255
PASS data[1] is 0
PASS data[2] is 0
PASS data[3] is around 31

PASS successfullyParsed is true

TEST COMPLETE
